Update xgl from commit: 0c20ef79
* Implement MESH_SHADER
* Bump LLPC client version to 55
* Fix API hash and fake memory leak
* [Navi2] Shader tuning for WolfensteinII
* Add src stage in case image transitions out of undefined layout
* Update PAL Version in XGL to 770
* Update Khronos Vulkan Headers to 1.3.232
* Initial changes for VK_EXT_extended_dynamic_state3
* Fix the assert in AllocateShaderOutput
* [Navi22] Shader tuning for WolfYB
* Move internal RT shader configuration from module option to pipeline option
* Fix new version CTS failure in dEQP-VK.pipeline.pipeline_library.graphics_library.misc.non_graphics.*
* Disable skip dst cache inv by default for gfx10+
* ASTC/ETC2 GPU Decoder 3D texture support
* Avoid some RP transitions being skipped
* Switch back to exposing physical heap size
* Support ldsSpillLimitDwords tuning option
* Shader tuning for WolfensteinYB
* [Navi2x] Quake2RTX LLPC shader tuning
* Fix DescriptorUpdateTemplate
* LLPC shader tuning for Quake2RTX
* Fix new version CTS failure in dEQP-VK.ray_tracing_pipeline.acceleration_structures.query_pool_results.cpu*
* VK_EXT_device_fault - Driver Implementation
* Fix handling of WgpModeCu
* VK_EXT_vertex_input_dynamic_state -  Driver Implementation
* Disable anisotropy filter for Battlefield 1
* [cmake] cleanup BUILD_XLIB_XRANDR_SUPPORT

+++ icd/api/include/vk_compute_pipeline.h
@@ -98,14 +98,14 @@ protected:
         bool                                 hasRayTracing,
         uint32_t                             dispatchRaysUserDataOffset,
 #endif
-        uint32_t                             staticStateMask,
+        uint64_t                             staticStateMask,
         uint64_t                             apiHash);
 
     // Converted creation info parameters of the Vulkan compute pipeline
     struct CreateInfo
     {
         ImmedInfo                              immedInfo;
-        uint32_t                               staticStateMask;
+        uint64_t                               staticStateMask;
         Pal::ComputePipelineCreateInfo         pipeline;
         const PipelineLayout*                  pLayout;
     };
